added Client project and code to get proxy
